Optimize data transfer between backend and frontend
