+++ include/python_api.cuh
@@ -3,7 +3,7 @@
 // Description:  API for Python bindings of C++/CUDA
 // Authors:      Luong-Ha Nguyen & James-A. Goulet
 // Created:      October 19, 2022
-// Updated:      November 04, 2022
+// Updated:      December 03, 2022
 // Contact:      luongha.nguyen@gmail.com & james.goulet@polymtl.ca
 // Copyright (c) 2022 Luong-Ha Nguyen & James-A. Goulet. Some rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
@@ -14,11 +14,51 @@
 
 #include <memory>
 
+#include "common.h"
+#include "cost.h"
+#include "dataloader.h"
 #include "struct_var.h"
 #include "tagi_network.cuh"
 #include "tagi_network_base.h"
 #include "tagi_network_cpu.h"
-#include "utility_wrapper.h"
+
+class UtilityWrapper {
+   public:
+    UtilityWrapper();
+    ~UtilityWrapper();
+    std::tuple<std::vector<float>, std::vector<int>, int> label_to_obs_wrapper(
+        std::vector<int> &labels, int num_classes);
+
+    std::tuple<pybind11::array_t<float>, pybind11::array_t<int>>
+    load_mnist_dataset_wrapper(std::string &image_file, std::string &label_file,
+                               int num);
+
+    std::tuple<pybind11::array_t<float>, pybind11::array_t<int>>
+    load_cifar_dataset_wrapper(std::string &image_file, int num);
+
+    std::tuple<std::vector<int>, std::vector<float>> get_labels_wrapper(
+        std::vector<float> &mz, std::vector<float> &Sz, HrSoftmax &hs,
+        int num_classes, int B);
+
+    HrSoftmax hierarchical_softmax_wrapper(int num_classes);
+
+    std::vector<float> obs_to_label_prob_wrapper(std::vector<float> &mz,
+                                                 std::vector<float> &Sz,
+                                                 HrSoftmax &hs,
+                                                 int num_classes);
+    std::tuple<std::vector<int>, std::vector<float>> get_error_wrapper(
+        std::vector<float> &mz, std::vector<float> &Sz,
+        std::vector<int> &labels, HrSoftmax &hs, int n_classes, int B);
+
+    std::tuple<std::vector<float>, std::vector<float>>
+    create_rolling_window_wrapper(std::vector<float> &data,
+                                  std::vector<int> &output_col,
+                                  int input_seq_len, int output_seq_len,
+                                  int num_features, int stride);
+
+    std::vector<float> get_upper_triu_cov_wrapper(int batch_size, int num_data,
+                                                  float &sigma);
+};
 
 class NetworkWrapper
 /* Python wrapper for C+/CUDA backend
@@ -45,27 +85,29 @@ Attributes:
 
     void param_feed_backward_wrapper();
 
-    std::tuple<std::vector<float>, std::vector<float>>
+    std::tuple<pybind11::array_t<float>, pybind11::array_t<float>>
     get_network_outputs_wrapper();
 
-    std::tuple<std::vector<float>, std::vector<float>>
+    std::tuple<pybind11::array_t<float>, pybind11::array_t<float>>
     get_network_prediction_wrapper();
 
-    std::tuple<std::vector<float>, std::vector<float>, std::vector<float>,
-               std::vector<float>, std::vector<float>>
+    std::tuple<pybind11::array_t<float>, pybind11::array_t<float>,
+               pybind11::array_t<float>, pybind11::array_t<float>,
+               pybind11::array_t<float>>
     get_all_network_outputs_wrapper();
 
-    std::tuple<std::vector<float>, std::vector<float>, std::vector<float>,
-               std::vector<float>, std::vector<float>>
+    std::tuple<pybind11::array_t<float>, pybind11::array_t<float>,
+               pybind11::array_t<float>, pybind11::array_t<float>,
+               pybind11::array_t<float>>
     get_all_network_inputs_wrapper();
 
-    std::tuple<std::vector<float>, std::vector<float>> get_derivative_wrapper(
-        int layer);
+    std::tuple<pybind11::array_t<float>, pybind11::array_t<float>>
+    get_derivative_wrapper(int layer);
 
-    std::tuple<std::vector<float>, std::vector<float>>
+    std::tuple<pybind11::array_t<float>, pybind11::array_t<float>>
     get_inovation_mean_var_wrapper(int layer);
 
-    std::tuple<std::vector<float>, std::vector<float>>
+    std::tuple<pybind11::array_t<float>, pybind11::array_t<float>>
     get_state_delta_mean_var_wrapper();
 
     void set_parameters_wrapper(Param &init_theta);
