+++ pytagi/tagi_network.py
@@ -3,7 +3,7 @@
 # Description:  Python frontend for TAGI network
 # Authors:      Luong-Ha Nguyen & James-A. Goulet
 # Created:      October 13, 2022
-# Updated:      November 12, 2022
+# Updated:      December 03, 2022
 # Contact:      luongha.nguyen@gmail.com & james.goulet@polymtl.ca
 # Copyright (c) 2022 Luong-Ha Nguyen & James-A. Goulet. Some rights reserved.
 ###############################################################################
@@ -230,7 +230,7 @@ class TagiNetwork:
         """
         ma, va = self.network.get_network_outputs_wrapper()
 
-        return np.array(ma), np.array(va)
+        return ma, va
 
     def get_network_predictions(self) -> Tuple[np.ndarray, np.ndarray]:
         """Get distribution of the predictions
@@ -241,7 +241,7 @@ class TagiNetwork:
         """
         m_pred, v_pred = self.network.get_network_prediction_wrapper()
 
-        return np.array(m_pred), np.array(v_pred)
+        return m_pred, v_pred
 
     def get_all_network_outputs(
         self
@@ -257,8 +257,7 @@ class TagiNetwork:
         """
         ma, va, mz, vz, jcb = self.network.get_all_network_outputs_wrapper()
 
-        return (np.array(ma), np.array(va), np.array(mz), np.array(vz),
-                np.array(jcb))
+        return (ma, va, mz, vz, jcb)
 
     def get_all_network_inputs(
         self
@@ -274,8 +273,7 @@ class TagiNetwork:
         """
         ma, va, mz, vz, jcb = self.network.get_all_network_inputs_wrapper()
 
-        return (np.array(ma), np.array(va), np.array(mz), np.array(vz),
-                np.array(jcb))
+        return (ma, va, mz, vz, jcb)
 
     def get_derivatives(self, layer: int = 0) -> Tuple[np.ndarray, np.ndarray]:
         """ Compute derivatives of the output layer w.r.t a given layer using TAGI
@@ -303,7 +301,7 @@ class TagiNetwork:
 
         delta_m, delta_v = self.network.get_inovation_mean_var_wrapper(layer)
 
-        return np.array(delta_m), np.array(delta_v)
+        return delta_m, delta_v
 
     def get_state_delta_mean_var(self) -> None:
         """Get updating quatities for the first layer
@@ -316,7 +314,7 @@ class TagiNetwork:
         """
         delta_mz, delta_vz = self.network.get_state_delta_mean_var_wrapper()
 
-        return np.array(delta_mz), np.array(delta_vz)
+        return delta_mz, delta_vz
 
     def set_parameters(self, param: Param) -> None:
         """Set parameter values to network"""
